Extract cli-engine
