+++ packages/cli/libs/commands/run/complete.command.ts
@@ -1,25 +1,17 @@
 import { readConfig } from "../../data-access/config.dao";
-import {
-  createEngine,
-  StepResult,
-  Engine,
-  State,
-  Environment,
-  isSuccess,
-} from "@jyfti/engine";
-import { last, mergeMap, tap, catchError, takeWhile } from "rxjs/operators";
-import { from, OperatorFunction, of } from "rxjs";
+import { createEngine, Environment } from "@jyfti/engine";
+import { catchError } from "rxjs/operators";
+import { of } from "rxjs";
 import { promptWorkflow } from "../../inquirer.service";
-import { printOutput, printStepResult } from "../../print.service";
 import {
   readWorkflowOrTerminate,
   readWorkflowNamesOrTerminate,
 } from "../../data-access/workflow.dao";
-import { writeState, readStateOrTerminate } from "../../data-access/state.dao";
+import { readStateOrTerminate } from "../../data-access/state.dao";
 import { readEnvironmentOrTerminate } from "../../data-access/environment.dao";
 import { validateEnvironmentOrTerminate } from "../../validator";
-import { Config } from "../../types/config";
 import { mergeEnvironments } from "../../data-access/environment.util";
+import { runToCompletion } from "../../cli-engine";
 
 export async function complete(
   name?: string,
@@ -41,37 +33,6 @@ export async function complete(
       cmd?.envVar || {},
     ]);
     validateEnvironmentOrTerminate(workflow, environment);
-    const engine = createEngine(workflow, environment, config.outRoot);
-    await engine
-      .complete(state)
-      .pipe(
-        process(engine, config, name, state),
-        catchError((err) => of(console.error("Unexpected Jyfti error", err)))
-      )
-      .toPromise();
+    await runToCompletion(workflow, environment, config, name, state);
   }
-}
-
-function process(
-  engine: Engine,
-  config: Config,
-  name: string,
-  state: State
-): OperatorFunction<StepResult, void> {
-  return (stepResult$) =>
-    stepResult$.pipe(
-      tap((stepResult) => console.log(printStepResult(stepResult))),
-      takeWhile(isSuccess, true),
-      engine.transitionFrom(state),
-      last(),
-      tap((state) => {
-        if (!state.error) {
-          const output = engine.getOutput(state);
-          if (output) {
-            console.log(printOutput(output));
-          }
-        }
-      }),
-      mergeMap((state) => from(writeState(config, name, state)))
-    );
 }
